Agrupando testes funcionais com pytest markers
