+++ .env-example
@@ -2,7 +2,11 @@
 PER_PAGE = 9
 
 # Django secret key
-SECRET_KEY = 'CHANGE-ME'
+SECRET_KEY = 'django-insecure-$$#oyl$2oqhc68t-u3wtvw%j5vmon25wpg((arphu!hl3@dq$3'
 
 # 0 = False - 1 = True
-DEBUG = 0
+DEBUG = 1
+
+# 0 = False - 1 = True
+
+SELENIUM_HEADLESS = 1
