+++ pytest.ini
@@ -8,3 +8,4 @@ addopts =
 markers =
   slow: Run tests that are slow
   fast: Run fast tests
+  function_test: Run tests that are selenium based
