+++ tests/function_tests/recipes/base.py
@@ -16,10 +16,3 @@ class RecipeBaseFunctionalTest(StaticLiveServerTestCase):
 
     def sleep(self, seconds=5):
         time.sleep(seconds)
-
-
-class RecipeHomePageFunctionalTest(RecipeBaseFunctionalTest):
-    def test_recipe_home_page_without_recipes_not_found_message(self):
-        self.browser.get(self.live_server_url)
-        body = self.browser.find_element(By.TAG_NAME, 'body')
-        self.assertIn('No recipes found here 🥲', body.text)
