+++ tests/function_tests/recipes/test_recipe_home_page.py
@@ -1,0 +1,12 @@
+import pytest
+from selenium.webdriver.common.by import By
+
+from .base import RecipeBaseFunctionalTest
+
+
+@pytest.mark.function_test
+class RecipeHomePageFunctionalTest(RecipeBaseFunctionalTest):
+    def test_recipe_home_page_without_recipes_not_found_message(self):
+        self.browser.get(self.live_server_url)
+        body = self.browser.find_element(By.TAG_NAME, 'body')
+        self.assertIn('No recipes found here 🥲', body.text)
